BookStore Functionlity class add useing  ES6.
