+++ source/js/index.js
@@ -3,6 +3,7 @@
 /* eslint-disable no-trailing-spaces */
 
 import * as variables from '../modules/values.js';
+import BookStore from '../modules/BookStore.js';
 
 variables.bookContainer;
 variables.form;
@@ -13,58 +14,6 @@ variables.addBookNav;
 variables.contactNav;
 variables.months;
 variables.dateDisplay;
-
-let bookList = [];
-class AwesomeBook {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-  }
-
-  // Remove Button
-  removeBook(btnAdderss, bookArray, btnIndex, parentDiv) {
-    btnAdderss.addEventListener('click', () => {
-      bookArray.splice(btnIndex, 1);
-      localStorage.setItem('book', JSON.stringify(bookArray));
-      parentDiv.remove();
-      window.location.reload();
-    });
-  }
-
-  // Adding Books
-  addBook() {
-    const bookData = JSON.parse(localStorage.getItem('book'));
-    bookData.forEach((item, index) => {
-      const bookDiv = document.createElement('div');
-      const bookItem = document.createElement('p');
-      const removeButton = document.createElement('button');
-      bookItem.appendChild(document.createTextNode(`${item.title} by ${item.author}`));
-      removeButton.setAttribute('class', 'remove');
-      bookItem.setAttribute('class', 'book');
-      bookDiv.setAttribute('class', 'book-card');
-      removeButton.appendChild(document.createTextNode('Remove'));
-      // Remove Button Logic
-      this.removeBook(removeButton, bookData, index, bookDiv);
-      bookDiv.append(bookItem, removeButton);
-      bookContainer.appendChild(bookDiv);
-    });
-  }
-
-  // Displaying Books
-  displayBooks() {
-    const book = {
-      title: this.title.value,
-      author: this.author.value,
-    };
-    bookList.push(book);
-    localStorage.setItem('book', JSON.stringify(bookList));
-    bookContainer.innerHTML = '';
-    this.addBook();
-    bookTitle.value = '';
-    bookAuthor.value = '';
-  }
-}
-
 function showDate() {
   const dateObj = new Date();
   const month = variables.months[dateObj.getMonth()];
@@ -82,10 +31,10 @@ function showDate() {
   variables.dateDisplay.innerText = `${month} ${day}th ${year}, ${hour}:${minutes}:${secnod}${amPM}`;
 }
 
-const bookObj = new AwesomeBook(variables.bookTitle, variables.bookAuthor);
+const bookObj = new BookStore(variables.bookTitle, variables.bookAuthor);
 
 if (localStorage.getItem('book') !== null) {
-  bookList = JSON.parse(localStorage.getItem('book'));
+  bookObj.setBookList();
   bookObj.addBook();
 }
